// src/app/layout/manage/model/prov-ext/provext.ts
export class ProvExt{
    nid_external: number = 0;
    edescription: string = "";
    nid_categoryhours: number = 0;
    namecategory: string = "";
    costohora: number = 0;
    sactive: string = ""; 
    dregister: string = "";
    nuserregister: number = 0;
    dupdate: string = "";
    nuserupdate: number = 0;
}